+++ internal/controller/nats/unit_test.go
@@ -2,7 +2,6 @@ package nats
 
 import (
 	"context"
-	"github.com/kyma-project/nats-manager/internal/metrics"
 	"testing"
 
 	nmapiv1alpha1 "github.com/kyma-project/nats-manager/api/v1alpha1"
@@ -10,6 +9,7 @@ import (
 	nmkchartmocks "github.com/kyma-project/nats-manager/pkg/k8s/chart/mocks"
 	nmkmocks "github.com/kyma-project/nats-manager/pkg/k8s/mocks"
 	nmmgrmocks "github.com/kyma-project/nats-manager/pkg/manager/mocks"
+	"github.com/kyma-project/nats-manager/pkg/metrics"
 	"github.com/kyma-project/nats-manager/testutils"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
